Assignment operator is now :=
This is only for assignment and not in LET, for the moment.
Moved code around.
The prelude is now loaded from a prelude.mini file, if it exists.
Commented code from the disabled interpreter and renamed CREPL to
REPL.
Fixed RANGE so it's comfortable to invoke as RANGE(a, b).