Reworked the help system
HELP() without parameters prints a list of unburied symbols.
HELP(symb) with a string parameter prints the documentation for that
symbol, if it exists.
DOC!(symb, text) sets the documentation for a symbol in
MINILANG-RUNTIME.
BURY(symb) buries a symbol, so it won't be listed by HELP().
UNBURY(symb) makes a symbol appear in the HELP() list.