Reworked the trail, still unfinished
The trail is now a list of nodes.

The idea is that each node carries a primitive, color, line width and
any other necessary information, and knows how to draw itself.

For the moment, the VERTICES slot is hijacked to contain exactly the
same kind of line information as before, and the nodes are created
otherwise empty.