Refactored the trail
*TRAIL* now holds an instance of the class TRAIL, which holds a list
of NODEs.

Each node knows what kind of primitive it represents.  LINES-NODE now
represents a list of vertices for line strips.  The nodes now hold
just a list of vertices. The drawing loop simply sends all the
vertices it contains, wrapped by a WITH-PRIMITIVE call.

When adding a new node to the trail, the top one is popped if it's
empty.

Added primitive SHAPE, to switch between kinds of node. Right now it
knows about lines and triangles nodes.

Added an optional A (alpha) argument to PAPER, defaults to 1.
Removed the kludge that was SAVE-DISPLAY.